spi: dw: add cs-gpio support to the Designware SPI controller
Introduce support to the Designware SPI controller for CS signals
that are connected to GPIO pins.  This includes support for active
high CS signals (which is not available with the builtin lines).

GPIO controlled CS signals can be used to either extend the set of
available chip select lines beyond the SPI controller's builtin set,
or to use CS lines that are under software control for those cases
where the SPI controller's builtin hardware control is not desired.
Test setups suggest that the SPI controller's builtin CS signals
are not under software control, and may have undesired behaviour
for SPI messages that consist of multiple transfers, or when the
CPU won't feed new data to the SPI controller fast enough for other
arbitrary reasons (caller requested delays, diagnostics, interrupt
load, high wire speed, etc).  DMA support does not solve the issue
for multi-transfer SPI messages or delays, which are provided by
the generic API.

Because of a limitation in either the hardware or in the existing
software implementation, SPI transfers stall the machine when their
associated chip select is none of the SPI controller's builtin CS
lines.  To work around this constraint, this implementation asserts
the last of the SPI controller's builtin CS lines for slaves that
reside on CS numbers beyond the builtin CS lines.  This occurs in
addition to controlling the GPIO connected CS line, it is done to
unbreak the communication in the first place, and the internal
CS signal can get ignored for those slaves, because only the GPIO
CS signal matters.  The workaround is considered to be a non-issue,
because the cs-gpios feature expands the available set of CS lines
in nearly unlimited ways, losing one potential line is not a blocker.