input: lnt-ti1: introduce common logic for LNT Touchinterface 1
Introduce the "core" driver which implements logic that is common for
all LNT TI1 variants.  This driver implements the sequence of commands
with their requests and responses, interprets exchanged packet data and
derives input events.

Additional drivers are required which implement the packet transport.
These are in subsequent commits.

The driver supports multi touch mode with up to ten fingers in the
LNT Touchinterface 1 core driver.  This feature is off by default,
and can get enabled by means of the 'multi' module option.