input: lnt-ti1: add SPI transport for LNT Touchinterface 1
Implement SPI communication for the LNT Touchinterface 1 touchscreen
controller.  This "transport" driver attaches to SPI slaves, and
registers with the common logic for LNT TI1 in the core driver.

A specific firmware feature of the SPI transport does transpire into the
common logic of the LNT Touchinterface 1 driver.  Whip up support for
the "deferred response" communication pattern, where transferring one
request will yield the response to a previously communicated request,
and the request that was communicated in that very transfer will get
responded to in a subsequent transfer.