spi: dw: eliminate alloc related WARNING in the Designware SPI master
Only probe() is supposed to devm_kzalloc() resources, setup() should
not.  Do the allocation and release of the 'chip' data in the setup()
and cleanup() callbacks by means of regular kzalloc() and kfree() calls,
instead of their devm variant.

This change eliminates a scary WARNING() which gets emitted by dd/base.c
in the case where unexpected allocations were detected.

The previous implementation had an additional bug in that devm_kzalloc()
was called in setup(), yet an explicit kfree() was called in cleanup().