HACK: rtc: m41t80: Prevent overflow of 32bit timer_t systems
The driver assumes that no dates before the year 2000 will be used and
thus silently changes e.g. 1990 to 2090 when writing to the RTC.  In
an unintitalized system without timesource that comes up as Jan 1,
1970 this will result in the date Jan 1, 2070 to be written.  On a
system with 32 bit timer_t this will overflow the internal time
keeping on the next read.

Fix this by truncating 19xx years to 2000 on write path and years
later than 2038 also to 2k on read path.

Signed-off-by: Detlev Zundel <detlev.zundel@ebv.com>